Rename BodyFragmentNew to BodyFragment
