+++ app/src/main/java/app/insti/fragment/BodyFragment.java
@@ -5,15 +5,13 @@ import android.animation.Animator;
 import android.animation.AnimatorListenerAdapter;
 import android.animation.AnimatorSet;
 import android.animation.ObjectAnimator;
-import android.content.Intent;
 import android.graphics.Color;
 import android.graphics.Point;
 import android.graphics.Rect;
-import android.net.Uri;
 import android.os.Bundle;
+import android.support.annotation.NonNull;
 import android.support.design.widget.FloatingActionButton;
 import android.support.v4.app.Fragment;
-import android.support.v4.widget.NestedScrollView;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
@@ -22,11 +20,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.DecelerateInterpolator;
-import android.widget.Button;
-import android.widget.ImageButton;
 import android.widget.ImageView;
-import android.widget.TextView;
-import android.widget.Toast;
 
 import com.google.gson.Gson;
 import com.squareup.picasso.Picasso;
@@ -36,21 +30,19 @@ import java.util.List;
 
 import app.insti.Constants;
 import app.insti.R;
-import app.insti.ShareURLMaker;
 import app.insti.Utils;
 import app.insti.activity.MainActivity;
-import app.insti.adapter.BodyAdapter;
-import app.insti.adapter.FeedAdapter;
-import app.insti.adapter.UserAdapter;
+import app.insti.adapter.GenericAdapter;
 import app.insti.api.RetrofitInterface;
 import app.insti.api.model.Body;
-import app.insti.api.model.Event;
 import app.insti.api.model.Role;
 import app.insti.api.model.User;
+import app.insti.interfaces.CardInterface;
+import app.insti.utils.BodyHeadCard;
+import app.insti.utils.TitleCard;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
-import ru.noties.markwon.Markwon;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -58,10 +50,32 @@ import ru.noties.markwon.Markwon;
  * create an instance of this fragment.
  */
 public class BodyFragment extends BackHandledFragment implements TransitionTargetFragment {
+    // TODO: Rename parameter arguments, choose names that match
+    // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
+
+
     String TAG = "BodyFragment";
+    // TODO: Rename and change types of parameters
     private Body min_body;
+    private SwipeRefreshLayout bodySwipeRefreshLayout;
+
+
+    // Hold a reference to the current animator,
+    // so that it can be canceled mid-way.
+    private Animator mCurrentAnimator;
+
+    // The system "short" animation time duration, in milliseconds. This
+    // duration is ideal for subtle animations or animations that occur
+    // very frequently.
+    private int mShortAnimationDuration;
+    private boolean zoomMode;
+    private ImageView expandedImageView;
+    private Rect startBounds;
+    private float startScaleFinal;
+    private ImageView bodyPicture;
     private Body body;
     private boolean bodyDisplayed = false;
+    private boolean transitionEnded = false;
 
     public BodyFragment() {
         // Required empty public constructor
@@ -85,6 +99,18 @@ public class BodyFragment extends BackHandledFragment implements TransitionTarge
 
     @Override
     public boolean onBackPressed() {
+        if (zoomMode) {
+            zoomOut(expandedImageView, startBounds, startScaleFinal, bodyPicture);
+            zoomMode = false;
+
+            /* Show fab if the user has access */
+            if (((MainActivity) getActivity()).editBodyAccess(body)) {
+                final FloatingActionButton fab = getView().findViewById(R.id.edit_fab);
+                fab.show();
+            }
+
+            return true;
+        }
         return false;
     }
 
@@ -99,6 +125,9 @@ public class BodyFragment extends BackHandledFragment implements TransitionTarge
     @Override
     public void transitionEnd() {
         if (getActivity() == null || getView() == null) return;
+        bodyPicture = (ImageView) getView().findViewById(R.id.body_picture);
+        Utils.loadImageWithPlaceholder(bodyPicture, body.getBodyImageURL());
+        transitionEnded = true;
     }
 
     @Override
@@ -107,10 +136,25 @@ public class BodyFragment extends BackHandledFragment implements TransitionTarge
 
         /* Initialize */
         bodyDisplayed = false;
-        body = min_body;
+
+        int index = Utils.bodyCache.indexOf(min_body);
+        if (index < 0) {
+            body = min_body;
+            updateBody();
+        } else {
+            body = Utils.bodyCache.get(index);
+        }
+
         displayBody();
 
-        updateBody();
+        bodySwipeRefreshLayout = getActivity().findViewById(R.id.body_swipe_refresh_layout);
+        bodySwipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
+            @Override
+            public void onRefresh() {
+                bodyDisplayed = false;
+                updateBody();
+            }
+        });
 
         Toolbar toolbar = getActivity().findViewById(R.id.toolbar);
         toolbar.setTitle(min_body.getBodyName());
@@ -128,17 +172,19 @@ public class BodyFragment extends BackHandledFragment implements TransitionTarge
             public void onResponse(Call<Body> call, Response<Body> response) {
                 if (response.isSuccessful()) {
                     Body bodyResponse = response.body();
+                    Utils.bodyCache.updateCache(response.body());
 
                     if (!bodyDisplayed) {
                         body = bodyResponse;
                         displayBody();
                     }
+                    bodySwipeRefreshLayout.setRefreshing(false);
                 }
             }
 
             @Override
             public void onFailure(Call<Body> call, Throwable t) {
-                // Network Error
+                bodySwipeRefreshLayout.setRefreshing(false);
             }
         });
     }
@@ -146,87 +192,230 @@ public class BodyFragment extends BackHandledFragment implements TransitionTarge
     private void displayBody() {
         /* Skip if we're already destroyed */
         if (getActivity() == null || getView() == null) return;
-        if (!body.equals(min_body)) bodyDisplayed = true;
+        if (!(body == min_body)) bodyDisplayed = true;
 
-        TextView bodyName = (TextView) getView().findViewById(R.id.body_name);
-        TextView bodySubtitle = getView().findViewById(R.id.body_subtitle);
-        TextView bodyDescription = (TextView) getView().findViewById(R.id.body_description);
-        ImageButton webBodyButton = getActivity().findViewById(R.id.web_body_button);
-        ImageButton shareBodyButton = getActivity().findViewById(R.id.share_body_button);
-        final Button followButton = getActivity().findViewById(R.id.follow_button);
+        bodyPicture = (ImageView) getActivity().findViewById(R.id.body_picture);
 
-        /* Set body information */
-        bodyName.setText(body.getBodyName());
-        bodySubtitle.setText(body.getBodyShortDescription());
+        /* Load only low res image if transition is not completed */
+        if (transitionEnded) {
+            Utils.loadImageWithPlaceholder(bodyPicture, body.getBodyImageURL());
+        } else {
+            Picasso.get().load(Utils.resizeImageUrl(body.getBodyImageURL())).into(bodyPicture);
+        }
 
-        /* Return if it's a min body */
-        if (body.getBodyDescription() == null) {
+        /* Skip for min body */
+        if (body == min_body) {
             return;
         }
 
-        Markwon.setMarkdown(bodyDescription, body.getBodyDescription());
-
-        /* Check if user is already following
-         * Initialize follow button */
-        followButton.setBackgroundColor(getResources().getColor(body.getBodyUserFollows() ? R.color.colorAccent : R.color.colorWhite));
-        followButton.setText(EventFragment.getCountBadgeSpannable("FOLLOW", body.getBodyFollowersCount()));
-
-        followButton.setOnClickListener(new View.OnClickListener() {
+        bodyPicture.setOnClickListener(new View.OnClickListener() {
             @Override
-            public void onClick(View view) {
-                RetrofitInterface retrofitInterface = Utils.getRetrofitInterface();
-                retrofitInterface.updateBodyFollowing(Utils.getSessionIDHeader(), body.getBodyID(), body.getBodyUserFollows() ? 0 : 1).enqueue(new Callback<Void>() {
-                    @Override
-                    public void onResponse(Call<Void> call, Response<Void> response) {
-                        if (response.isSuccessful()) {
-                            body.setBodyUserFollows(!body.getBodyUserFollows());
-                            body.setBodyFollowersCount(body.getBodyUserFollows()? body.getBodyFollowersCount()+1:body.getBodyFollowersCount()-1);
-                            followButton.setBackgroundColor(getResources().getColor(body.getBodyUserFollows() ? R.color.colorAccent : R.color.colorWhite));
-                            followButton.setText(EventFragment.getCountBadgeSpannable("FOLLOW", body.getBodyFollowersCount()));
-                        }
-                    }
-
-                    @Override
-                    public void onFailure(Call<Void> call, Throwable t) {
-                        Toast.makeText(getContext(), "Network Error", Toast.LENGTH_LONG).show();
-                    }
-                });
+            public void onClick(View v) {
+                zoomImageFromThumb(bodyPicture);
+                final FloatingActionButton fab = getView().findViewById(R.id.edit_fab);
+                fab.hide();
             }
         });
+        mShortAnimationDuration = getResources().getInteger(android.R.integer.config_shortAnimTime);
 
-        /* Initialize web button */
-        if (body.getBodyWebsiteURL() != null && !body.getBodyWebsiteURL().isEmpty()) {
-            webBodyButton.setVisibility(View.VISIBLE);
-            webBodyButton.setOnClickListener(new View.OnClickListener() {
-                String bodywebURL = body.getBodyWebsiteURL();
+        final List<Role> roles = body.getBodyRoles();
+        final List<User> users = new ArrayList<>();
+        for (Role role : roles) {
+            if (role.getRoleUsersDetail() != null) {
+                for (User user : role.getRoleUsersDetail()) {
+                    user.setCurrentRole(role.getRoleName());
+                    users.add(user);
+                }
+            }
+        }
 
+        final List<CardInterface> cards = new ArrayList<>();
+        cards.add(new BodyHeadCard(body));
+        addWithTitleCard(cards,  body.getBodyEvents(), "Events");
+        addWithTitleCard(cards,  users, "People");
+        addWithTitleCard(cards,  body.getBodyChildren(), "Organizations");
+        addWithTitleCard(cards,  body.getBodyParents(), "Part of");
+
+        final RecyclerView recyclerView = (RecyclerView) getActivity().findViewById(R.id.body_recycler_view);
+        GenericAdapter genericAdapter =  new GenericAdapter(cards, this);
+        recyclerView.setAdapter(genericAdapter);
+        recyclerView.setLayoutManager(new LinearLayoutManager(getContext()));
+
+        getActivity().findViewById(R.id.loadingPanel).setVisibility(View.GONE);
+
+        /* Show update button if role */
+        if (((MainActivity) getActivity()).editBodyAccess(body)) {
+            final FloatingActionButton fab = getView().findViewById(R.id.edit_fab);
+            fab.show();
+            recyclerView.addOnScrollListener(new RecyclerView.OnScrollListener() {
                 @Override
-                public void onClick(View view) {
-                    Intent browserIntent = new Intent(Intent.ACTION_VIEW, Uri.parse(bodywebURL));
-                    startActivity(browserIntent);
+                public void onScrolled(@NonNull RecyclerView recyclerView, int dx, int dy) {
+                    super.onScrolled(recyclerView, dx, dy);
+                    if (dy > 0) fab.hide();
+                    else fab.show();
+                }
+            });
+
+            fab.setOnClickListener(new View.OnClickListener() {
+                @Override
+                public void onClick(View v) {
+                    AddEventFragment addEventFragment = new AddEventFragment();
+                    Bundle bundle = new Bundle();
+                    bundle.putString("bodyId", body.getBodyID());
+                    addEventFragment.setArguments(bundle);
+                    ((MainActivity) getActivity()).updateFragment(addEventFragment);
                 }
             });
         }
+    }
 
-        /* Initialize share button */
-        shareBodyButton.setOnClickListener(new View.OnClickListener() {
-            String shareUrl = ShareURLMaker.getBodyURL(body);
-
-            @Override
-            public void onClick(View view) {
-                Intent i = new Intent(Intent.ACTION_SEND);
-                i.setType("text/plain");
-                i.putExtra(Intent.EXTRA_SUBJECT, "Sharing URL");
-                i.putExtra(Intent.EXTRA_TEXT, shareUrl);
-                startActivity(Intent.createChooser(i, "Share URL"));
-            }
-        });
+    private <R extends CardInterface> void addWithTitleCard(List<CardInterface> cards, List<R> cardsToAdd, String title) {
+        if (cardsToAdd.size() == 0) return;
+        cards.add(new TitleCard(title));
+        cards.addAll(cardsToAdd);
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_body, container, false);
+        return inflater.inflate(R.layout.fragment_body_fragment_new, container, false);
+    }
+
+    private void zoomImageFromThumb(final ImageView thumbView) {
+        // If there's an animation in progress, cancel it
+        // immediately and proceed with this one.
+        if (mCurrentAnimator != null) {
+            mCurrentAnimator.cancel();
+        }
+
+        // Load the high-resolution "zoomed-in" image.
+        expandedImageView = (ImageView) getView().findViewById(
+                R.id.expanded_image_body);
+        expandedImageView.setImageDrawable(thumbView.getDrawable());
+
+        // Calculate the starting and ending bounds for the zoomed-in image.
+        // This step involves lots of math. Yay, math.
+        startBounds = new Rect();
+        final Rect finalBounds = new Rect();
+        final Point globalOffset = new Point();
+
+        // The start bounds are the global visible rectangle of the thumbnail,
+        // and the final bounds are the global visible rectangle of the container
+        // view. Also set the container view's offset as the origin for the
+        // bounds, since that's the origin for the positioning animation
+        // properties (X, Y).
+        thumbView.getGlobalVisibleRect(startBounds);
+        getActivity().findViewById(R.id.container_body)
+                .getGlobalVisibleRect(finalBounds, globalOffset);
+        startBounds.offset(-globalOffset.x, -globalOffset.y);
+        finalBounds.offset(-globalOffset.x, -globalOffset.y);
+
+        // Adjust the start bounds to be the same aspect ratio as the final
+        // bounds using the "center crop" technique. This prevents undesirable
+        // stretching during the animation. Also calculate the start scaling
+        // factor (the end scaling factor is always 1.0).
+        float startScale;
+        if ((float) finalBounds.width() / finalBounds.height()
+                > (float) startBounds.width() / startBounds.height()) {
+            // Extend start bounds horizontally
+            startScale = (float) startBounds.height() / finalBounds.height();
+            float startWidth = startScale * finalBounds.width();
+            float deltaWidth = (startWidth - startBounds.width()) / 2;
+            startBounds.left -= deltaWidth;
+            startBounds.right += deltaWidth;
+        } else {
+            // Extend start bounds vertically
+            startScale = (float) startBounds.width() / finalBounds.width();
+            float startHeight = startScale * finalBounds.height();
+            float deltaHeight = (startHeight - startBounds.height()) / 2;
+            startBounds.top -= deltaHeight;
+            startBounds.bottom += deltaHeight;
+        }
+
+        // Hide the thumbnail and show the zoomed-in view. When the animation
+        // begins, it will position the zoomed-in view in the place of the
+        // thumbnail.
+        thumbView.setAlpha(0f);
+        expandedImageView.setVisibility(View.VISIBLE);
+
+        // Set the pivot point for SCALE_X and SCALE_Y transformations
+        // to the top-left corner of the zoomed-in view (the default
+        // is the center of the view).
+        expandedImageView.setPivotX(0f);
+        expandedImageView.setPivotY(0f);
+
+        // Construct and run the parallel animation of the four translation and
+        // scale properties (X, Y, SCALE_X, and SCALE_Y).
+        AnimatorSet set = new AnimatorSet();
+        set
+                .play(ObjectAnimator.ofFloat(expandedImageView, View.X,
+                        startBounds.left, finalBounds.left))
+                .with(ObjectAnimator.ofFloat(expandedImageView, View.Y,
+                        startBounds.top, finalBounds.top))
+                .with(ObjectAnimator.ofFloat(expandedImageView, View.SCALE_X,
+                        startScale, 1f))
+                .with(ObjectAnimator.ofFloat(expandedImageView, View.SCALE_Y,
+                        startScale, 1f));
+        set.setDuration(mShortAnimationDuration);
+        set.setInterpolator(new DecelerateInterpolator());
+        set.addListener(new AnimatorListenerAdapter() {
+            @Override
+            public void onAnimationEnd(Animator animation) {
+                mCurrentAnimator = null;
+                expandedImageView.setBackgroundColor(Color.parseColor("#9E9E9E"));
+            }
+
+            @Override
+            public void onAnimationCancel(Animator animation) {
+                mCurrentAnimator = null;
+            }
+        });
+        set.start();
+        mCurrentAnimator = set;
+
+        startScaleFinal = startScale;
+        zoomMode = true;
+    }
+
+    private void zoomOut(final ImageView expandedImageView, Rect startBounds, float startScaleFinal, final View thumbView) {
+        expandedImageView.setBackgroundColor(0x00000000);
+        if (mCurrentAnimator != null) {
+            mCurrentAnimator.cancel();
+        }
+
+        // Animate the four positioning/sizing properties in parallel,
+        // back to their original values.
+        AnimatorSet set = new AnimatorSet();
+        set.play(ObjectAnimator
+                .ofFloat(expandedImageView, View.X, startBounds.left))
+                .with(ObjectAnimator
+                        .ofFloat(expandedImageView,
+                                View.Y, startBounds.top))
+                .with(ObjectAnimator
+                        .ofFloat(expandedImageView,
+                                View.SCALE_X, startScaleFinal))
+                .with(ObjectAnimator
+                        .ofFloat(expandedImageView,
+                                View.SCALE_Y, startScaleFinal));
+        set.setDuration(mShortAnimationDuration);
+        set.setInterpolator(new DecelerateInterpolator());
+        set.addListener(new AnimatorListenerAdapter() {
+            @Override
+            public void onAnimationEnd(Animator animation) {
+                thumbView.setAlpha(1f);
+                expandedImageView.setVisibility(View.GONE);
+                mCurrentAnimator = null;
+            }
+
+            @Override
+            public void onAnimationCancel(Animator animation) {
+                thumbView.setAlpha(1f);
+                expandedImageView.setVisibility(View.GONE);
+                mCurrentAnimator = null;
+            }
+        });
+        set.start();
+        mCurrentAnimator = set;
     }
 }
