+++ app/src/main/res/layout/fragment_body.xml
@@ -5,8 +5,7 @@
     android:id="@+id/container_body"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
-    android:orientation="vertical"
-    tools:context=".fragment.BodyFragment">
+    android:orientation="vertical">
 
     <android.support.v7.widget.CardView
         android:layout_width="match_parent"
