+++ app/src/main/res/layout/fragment_body_fragment_new.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".fragment.BodyFragmentNew"
+    tools:context=".fragment.BodyFragment"
     android:background="?attr/themeColor">
 
     <android.support.design.widget.CoordinatorLayout
